fixed logic behind start now and 1 min buffer and fixed alarm item view progress bars not properly positioned when alarm cancelled and reset

+++ app/src/main/java/com/example/crumb/Dao/ScheduleDao.kt
@@ -19,7 +19,7 @@ interface ScheduleDao {
     fun getSelectedAsLiveData(id : Long) : LiveData<Schedule>
 
     @Query("SELECT start_time from schedule_table where id = :id")
-    suspend fun getStartTime(id : Long) : Int
+    suspend fun getStartTime(id : Long) : Long
 
     @Query("SELECT description from schedule_table where id = :id")
     suspend fun getNotes(id : Long) : String
